requirements.txt
discord.py-self==2.4.0
protobuf==5.29.3
python-dotenv==1.0.1
